// src/models/FantasyDataPlayer.ts
export interface FantasyDataPlayer {
  rank: number;
  fullName: string;
  team: string;
  position: string;
  week: number;
  opponent: string;
  passingYds: number;
  passingTds: number;
  passingInts: number;
  rushingYds: number;
  rushingTds: number;
  receptions: number;
  receivingYds: number;
  receivingTds: number;
  sacks: number;
  interceptions: number;
  fumblesForced: number;
  fumblesRecovered: number;
  projectedPoints: number;
  averagePPG: number;
}